app/requirements.txt
django
django-filter
django-cors-headers
django-extensions
psycopg2-binary
djangorestframework
drf-spectacular
uwsgi
